MRI Analyzer/requirements.txt
google-generativeai
gradio
python-dotenv